Refactor: merge GameManager into main.ts
